requirements.txt
# Çekirdek
pydantic-settings>=2.2.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Zamanlama
APScheduler>=3.10.0

# Telegram: kanal okuma (kullanıcı oturumu) + onay botu
Telethon>=1.34.0
python-telegram-bot>=21.0

# OpenAI + X
openai>=1.30.0
tweepy>=4.14.0

# Yardımcı
httpx>=0.27.0

# Test (opsiyonel)
pytest>=8.0.0
